Adds refactoring
