+++ src/shapes.js
@@ -229,3 +229,22 @@ export const HistoryShape = PropTypes.shape({
   block: PropTypes.func.isRequired,
   replace: PropTypes.func.isRequired,
 });
+
+export const TopicConnectionShape = PropTypes.shape({
+  connectionId: PropTypes.string.isRequired,
+  isPrimary: PropTypes.bool,
+  paths: PropTypes.arrayOf(PropTypes.string),
+  targetId: PropTypes.string,
+  type: PropTypes.string,
+});
+
+export const TopicShape = PropTypes.shape({
+  connectionId: PropTypes.string.isRequired,
+  contentUri: PropTypes.string,
+  id: PropTypes.string,
+  name: PropTypes.string,
+  path: PropTypes.string,
+  primary: PropTypes.bool,
+  topicConnections: PropTypes.arrayOf(TopicConnectionShape),
+})
+
